Add an update button in settings
A button has been implemented that allows the user to update information about the selected user and service. Clicking it validates the selection and updates the database. Error messages have been added to handle cases where users or services are not found, as well as exception handling during the update.