Add Configuration Management and Update Installer
A new `envPath` variable has been added to the `InfoForm` class to store the path to the `config.env` configuration file, allowing the program to search for the file in the application's base directory. Multiple methods have been modified to use this new variable, and checks have been added to ensure the configuration file exists when launching the application.

In addition, sections have been added to the `computerinfo_installer.iss` installation file to define the installer configuration, including the application name, version, and file paths to include in the installation. Significant changes have also been made to `computerInfo_installer.exe`, including the addition of multiple lines of code that appear to be binary or configuration data, suggesting a modification or recompilation of the executable file.